requirements.txt
pandas>=1.5
numpy>=1.24
openpyxl>=3.0
python-dotenv>=1.0
